requirements.txt
Flask
Werkzeug
azure-storage-blob
python-dotenv